I18N: Bail early if an invalid text domain is passed to `load_textdomain()` et al.
Some plugins pass invalid values such as `null` instead of a string, which has never been supported by WordPress (no translations are loaded) and was technically undefined behavior. With the introduction of the new l10n library in #59656, which has stricter type hints, this could end up causing warnings or even fatal errors.

This change adds a deliberate short-circuit to `load_textdomain()` & co. to better handle such a case and document that it is not supported.

Props verygoode, swissspidy.
Fixes #60888.

git-svn-id: https://develop.svn.wordpress.org/trunk@57925 602fd350-edb4-49c9-b593-d223f7449a82